client/python/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mchat-client"
version = "0.1.0"
description = "MChat Python 客户端 SDK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "paho-mqtt>=2.0.0",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
]
[project.optional-dependencies]
dev = []

[tool.setuptools.packages.find]
where = ["src"]
